+++ R/ip_address.R
@@ -1,7 +1,7 @@
 #' @importFrom methods setOldClass
 methods::setOldClass(c("ip_address", "vctrs_vctr"))
 
-#' Class for storing IP addresses
+#' Vector of IP addresses
 #'
 #' @details
 #' An address in IPv4 space uses 32-bits. It is usually represented
@@ -26,7 +26,8 @@ methods::setOldClass(c("ip_address", "vctrs_vctr"))
 #' A special case is IPv4-mapped IPv6 addresses, which are returned in their
 #' dual representation (e.g. `::ffff:192.0.2.128`).
 #'
-#' This class also supports bitwise operations: `&` (AND), `|` (OR) and `!` (NOT).
+#' This class also supports bitwise operations: `!` (NOT), `&` (AND),
+#' `|` (OR) and `^` (XOR).
 #'
 #' @param x An `ip_address` vector
 #' @param ... Arguments to be passed to other methods
@@ -59,7 +60,10 @@ NULL
 #' ip_address("192.168.0.1") & ip_address("255.0.0.255")
 #'
 #' # bitwise OR
-#' ip_address("192.168.0.0") | ip_address("0.0.0.1")
+#' ip_address("192.168.0.0") | ip_address("255.0.0.255")
+#'
+#' # bitwise XOR
+#' ip_address("192.168.0.0") ^ ip_address("255.0.0.255")
 #' @rdname ip_address
 #' @export
 ip_address <- function(ip = character()) {