+++ src/IpAddressVector.h
@@ -67,9 +67,10 @@ public:
   /*---------------------*
    *  Bitwise operators  *
    *---------------------*/
+  IpAddressVector operator~() const;
   IpAddressVector operator&(const IpAddressVector &rhs) const;
   IpAddressVector operator|(const IpAddressVector &rhs) const;
-  IpAddressVector operator~() const;
+  IpAddressVector operator^(const IpAddressVector &rhs) const;
 
 
   /*-----------------------*