+++ src/wrapper.cpp
@@ -56,6 +56,12 @@ DataFrame compare_address_wrapper(List x) {
  *  Bitwise operators  *
  *---------------------*/
 // [[Rcpp::export]]
+List bitwise_not_wrapper(List addr1) {
+  IpAddressVector result = ~IpAddressVector(addr1);
+  return result.encodeR();
+}
+
+// [[Rcpp::export]]
 List bitwise_and_wrapper(List addr1, List addr2) {
   IpAddressVector result = IpAddressVector(addr1) & IpAddressVector(addr2);
   return result.encodeR();
@@ -68,8 +74,8 @@ List bitwise_or_wrapper(List addr1, List addr2) {
 }
 
 // [[Rcpp::export]]
-List bitwise_not_wrapper(List addr1) {
-  IpAddressVector result = ~IpAddressVector(addr1);
+List bitwise_xor_wrapper(List addr1, List addr2) {
+  IpAddressVector result = IpAddressVector(addr1) ^ IpAddressVector(addr2);
   return result.encodeR();
 }